Conflicts:
	src/Savegame/BattleUnit.cpp
	src/Savegame/SavedGame.h
	src/Savegame/Soldier.cpp

+++ src/Basescape/TransferItemsState.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2013 OpenXcom Developers.
+ * Copyright 2010-2014 OpenXcom Developers.
  *
  * This file is part of OpenXcom.
  *
@@ -780,7 +780,7 @@ enum TransferType TransferItemsState::getType(size_t selected) const
  * @param selected Currently selected item.
  * @return Index of the selected item.
  */
-int TransferItemsState::getItemIndex(size_t selected) const
+size_t TransferItemsState::getItemIndex(size_t selected) const
 {
 	return selected - _soldiers.size() - _crafts.size() - _hasSci - _hasEng;
 }
