+++ src/Engine/AdlibMusic.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2013 OpenXcom Developers.
+ * Copyright 2010-2014 OpenXcom Developers.
  *
  * This file is part of OpenXcom.
  *
@@ -47,7 +47,7 @@ public:
 	/// Loads music from the specified file.
 	void load(const std::string &filename);
 	/// Loads music from a chunk of memory.
-	void load(const void *data, size_t size);
+	void load(const void *data, int size);
 	/// Plays the music.
 	void play(int loop = -1) const;
 	/// Adlib music player.
