+++ src/Ruleset/SoldierNamePool.cpp
@@ -1,5 +1,5 @@
 /*
- * Copyright 2010-2013 OpenXcom Developers.
+ * Copyright 2010-2014 OpenXcom Developers.
  *
  * This file is part of OpenXcom.
  *
@@ -124,7 +124,7 @@ std::wstring SoldierNamePool::genName(SoldierGender *gender) const
  * @param numLooks The maximum index.
  * @return The index of the soldier's look.
  */
-int SoldierNamePool::genLook(size_t numLooks)
+size_t SoldierNamePool::genLook(size_t numLooks)
 {
 	int look = 0;
 	const int minimumChance = 2;	// minimum chance of a look being selected if it isn't enumerated. This ensures that looks MUST be zeroed to not appear.
